fix: 버스 정류소를 재검색하면 기존 정류소 마커는 제거
버스 정류소를 재검색해도 이전에 조회했던 정류소 마커들은 계속 유지된다.
그래서 조회할 수록 마커가 쌓이는 문제가 있다. 갯수에 그렇게 영향을 받는건
아니지만 굳이 유지할 필요도 없는 것 같다. 그러므로 정류소를 조회할 때
이전에 설정한 마커는 맵에서 제거하게 만든다. 마커 제거는 마커 인스턴스를
통해서 가능하므로 생성한 마커들을 저장해둬야 사용할 수 있다.
마커 목록 상태를 만들고 정류소 검색시마다 갱신하게 만든다.
그리고 이 마커 목록의 인스턴트들을 활용해서 맵에서 마커를 제거한다.
마커 목록 상태를 추가하면서 소스가 조금 복잡해져서 별도 훅으로 분리한다.
